requirements.txt
streamlit
fastapi
uvicorn
pyyaml
pydantic
python-multipart
requests
PyPDF2
matplotlib
reportlab